@@ -2,37 +2,18 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:intl/intl.dart';
+import 'package:medical_app/presentation/home/model/nurse_model.dart';
 import 'package:medical_app/presentation/home/widget/doctordetailscontainer.dart';
 import '../../../const/colors.dart';
 import '../../app/style/custom_button.dart';
-import 'doctorbooking.dart';
+import 'model/service.dart';
 
 class DoctorDetails extends StatefulWidget {
   const DoctorDetails({
     super.key,
-    required this.photoUrl,
-    required this.doctorName,
-    required this.category,
-    required this.rating,
-    required this.address,
-    required this.city,
-    required this.government,
-    required this.aboutUs,
-    required this.times,
-    required this.reservation,
-    required this.email,
+    required this.nurse,
   });
-  final String email;
-  final String photoUrl;
-  final String doctorName;
-  final String category;
-  final double rating;
-  final String address;
-  final String city;
-  final String government;
-  final String aboutUs;
-  final List<Timestamp> times;
-  final int reservation;
+  final Nurse nurse;
   static String id = 'DoctorDetails';
 
   @override
@@ -40,31 +21,58 @@ class DoctorDetails extends StatefulWidget {
 }
 
 class _DoctorDetailsState extends State<DoctorDetails> {
-  late DateTime _selectedDate;
-  late int checkin; // Define checkin variable
-  late List<DateTime> dateTimeList; // List to hold converted times
+  NurseService? _selectedService;
 
-  @override
-  void initState() {
-    super.initState();
-    _selectedDate = DateTime.now();
-    checkin = -1; // Initialize checkin variable
-    dateTimeList =
-        widget.times.map((timestamp) => timestamp.toDate()).toList();
+  void _bookAppointment() {
+    if (_selectedService == null) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(content: Text('Please select a service first')),
+      );
+      return;
+    }
+
+    final now = DateTime.now();
+    final formatted = DateFormat('yyyy-MM-dd HH:mm').format(now);
+
+    showDialog(
+      context: context,
+      builder: (_) => AlertDialog(
+        title: const Text('Appointment Details'),
+        content: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text('Nurse (AR): ${widget.nurse.nameAr}'),
+            Text('Nurse (EN): ${widget.nurse.nameEn}'),
+            Text('Phone: ${widget.nurse.phone}'),
+            const SizedBox(height: 8),
+            Text('Service (AR): ${_selectedService!.nameAr}'),
+            Text('Service (EN): ${_selectedService!.nameEn}'),
+            Text('Price: ${_selectedService!.price}'),
+            const SizedBox(height: 8),
+            Text('Date & Time: $formatted'),
+          ],
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: const Text('OK'),
+          ),
+        ],
+      ),
+    );
   }
 
   @override
   Widget build(BuildContext context) {
     double height = MediaQuery.of(context).size.height;
     double width = MediaQuery.of(context).size.width;
-    final dateFormat = DateFormat('h:mm a');
+    final services = widget.nurse.services ?? [];
 
     return Scaffold(
       appBar: AppBar(
         leading: GestureDetector(
-          onTap: () {
-            Navigator.pop(context);
-          },
+          onTap: () => Navigator.pop(context),
           child: const Icon(Icons.arrow_back_ios),
         ),
         centerTitle: true,
@@ -72,11 +80,11 @@ class _DoctorDetailsState extends State<DoctorDetails> {
           "Doctor Detail",
           style: GoogleFonts.inter(
             fontWeight: FontWeight.w600,
-            fontSize: MediaQuery.of(context).size.width < 500
-                ? MediaQuery.of(context).size.width * 0.0472
-                : MediaQuery.of(context).size.width < 800
-                ? MediaQuery.of(context).size.width * 0.05
-                : MediaQuery.of(context).size.width * 0.02,
+            fontSize: width < 500
+                ? width * 0.0472
+                : width < 800
+                ? width * 0.05
+                : width * 0.02,
           ),
         ),
       ),
@@ -87,135 +95,46 @@ class _DoctorDetailsState extends State<DoctorDetails> {
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
               DoctorDetailsContainer(
-                photoUrl: widget.photoUrl,
-                doctorName: widget.doctorName,
-                category: widget.category,
-                rating: widget.rating,
-                address: widget.address,
-
-
-                aboutDoctor: widget.aboutUs, city: widget.city, government:  widget.government,
+                photoUrl: widget.nurse.photo,
+                doctorName: widget.nurse.nameEn,
+                address: widget.nurse.cityEn,
               ),
-              const SizedBox(
-                height: 5,
-              ),
-              Container(
-                height: 336.7,
-                width: 341.1,
-                decoration: BoxDecoration(
-                  color: AppColor.kWhite,
-                  borderRadius: BorderRadius.circular(20),
-                ),
-                child: CalendarDatePicker(
-                  initialDate: DateTime.now(),
-                  firstDate: DateTime.now(),
-                  lastDate: DateTime(2130),
-                  onDateChanged: (dateTimeList) {
-                    setState(() {
-                      _selectedDate = dateTimeList;
-                      DateTime dateTime =
-                      DateTime.parse(_selectedDate.toString());
-                      String formattedDate =
-                      DateFormat('yyyy-MM-dd').format(dateTime);
-                    });
+              const SizedBox(height: 5),
+              SizedBox(
+                height:300,
+                child: ListView.builder(
+                  itemCount: services.length,
+                  itemBuilder: (context, index) {
+                    final svc = services[index];
+                    final selected = svc == _selectedService;
+                    return Card(
+                      color: selected ? AppColor.kPrimaryColor1.withOpacity(0.3) : null,
+                      child: ListTile(
+                        title: Text(svc.nameEn),
+                        subtitle: Text("${svc.price}"),
+                        onTap: () {
+                          setState(() => _selectedService = svc);
+                        },
+                      ),
+                    );
                   },
                 ),
               ),
-              const SizedBox(
-                height: 10,
-              ),
-              GridView.builder(
-                physics: const NeverScrollableScrollPhysics(),
-                shrinkWrap: true,
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 3,
-                  mainAxisSpacing: 15,
-                  crossAxisSpacing: 20,
-                  mainAxisExtent: 45,
+SizedBox(height: 100,),
+              Align(
+                alignment: Alignment.bottomCenter,
+                child: CustomButton(
+                  onTap: _bookAppointment,
+                  height: height * 0.0577,
+                  width: width * 0.7472,
+                  color: AppColor.kPrimaryColor1,
+                  borderRadius: BorderRadius.circular(width * 0.0779),
+                  border: Border.all(color: AppColor.kPrimaryColor1),
+                  text: 'Book Appointment',
+                  fontSize: width * 0.0365,
+                  colorText: AppColor.kWhite,
+                  fontWeight: FontWeight.w600,
                 ),
-                itemCount: dateTimeList.length,
-                itemBuilder: (BuildContext context, int index) {
-                  final formattedTime = dateFormat.format(dateTimeList[index]);
-
-                  return GestureDetector(
-                    onTap: () {
-                      setState(() {
-                        checkin = index;
-                      });
-                    },
-                    child: Container(
-                      width: 103,
-                      height: 37,
-                      decoration: BoxDecoration(
-                        color: checkin != index
-                            ? AppColor.kWhite
-                            : AppColor.kPrimaryColor1,
-                        border: Border.all(
-                          color: checkin != index
-                              ? AppColor.kWhite
-                              : AppColor.kPrimaryColor1,
-                        ),
-                        borderRadius:
-                        BorderRadius.circular(0.048661800486618 * width),
-                        boxShadow: const [
-                          BoxShadow(
-                            color: Color(0x3f000000),
-                            offset: Offset(0, 4),
-                            blurRadius: 2,
-                          ),
-                        ],
-                      ),
-                      child: Center(
-                        child: Text(
-                          formattedTime,
-                          style: GoogleFonts.inter(
-                            fontSize: 0.0389294403892944 * width,
-                            fontWeight: FontWeight.w600,
-                            color: checkin != index
-                                ? AppColor.kBlack
-                                : AppColor.kWhite,
-                          ),
-                        ),
-                      ),
-                    ),
-                  );
-                },
-              ),
-              SizedBox(
-                height: 0.0115340253748558 * height,
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  CustomButton(
-                    onTap: () {
-                      Navigator.of(context).push(MaterialPageRoute(
-                        builder: (context) => DoctorBooking(
-                          name: widget.doctorName,
-                          category: widget.category,
-                          rating: widget.rating,
-                          address: widget.address,
-                          photoUrl: widget.photoUrl,
-                          city: widget.city,
-                          government: widget.government,
-                          reservation: widget.reservation,
-                        ),
-                      ));
-                    },
-                    height: height * 0.0576701268742791,
-                    width: width * 0.7472019464720195,
-                    color: AppColor.kPrimaryColor1,
-                    borderRadius:
-                    BorderRadius.circular(0.0778588807785888 * width),
-                    border: Border.all(
-                      color: AppColor.kPrimaryColor1,
-                    ),
-                    text: 'Book Appointment',
-                    fontSize: width * 0.0364963503649635,
-                    colorText: AppColor.kWhite,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ],
               ),
             ],
           ),
