@@ -22,227 +22,265 @@ class MessageLookup extends MessageLookupByLibrary {
 
   final messages = _notInlinedMessages(_notInlinedMessages);
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
-        "ActiveIngredients":
-            MessageLookupByLibrary.simpleMessage("المواد الفعاله"),
-        "AnatomicalLocation":
-            MessageLookupByLibrary.simpleMessage("الموقع التشريحي"),
-        "Arteriesandbloodveins":
-            MessageLookupByLibrary.simpleMessage("الشرايين والأوردة الدموية"),
-        "Bones": MessageLookupByLibrary.simpleMessage("العظام"),
-        "Brain": MessageLookupByLibrary.simpleMessage("المخ"),
-        "Breast": MessageLookupByLibrary.simpleMessage("صدر"),
-        "Cancel": MessageLookupByLibrary.simpleMessage("الغاء"),
-        "Categories": MessageLookupByLibrary.simpleMessage("التخصصات"),
-        "Causes": MessageLookupByLibrary.simpleMessage("الاسباب"),
-        "Choose": MessageLookupByLibrary.simpleMessage("اختار"),
-        "Class": MessageLookupByLibrary.simpleMessage("العائله"),
-        "Classification": MessageLookupByLibrary.simpleMessage("التصنيف"),
-        "Company": MessageLookupByLibrary.simpleMessage("الشركة"),
-        "Consult_only_with_a_doctor_you_trust":
-            MessageLookupByLibrary.simpleMessage(
-                "استشر فقط الطبيب الذي تثق به"),
-        "Description": MessageLookupByLibrary.simpleMessage("الوصف"),
-        "Details": MessageLookupByLibrary.simpleMessage("التفاصيل"),
-        "DetailsAboutDiseases":
-            MessageLookupByLibrary.simpleMessage("تفاصيل عن الأمراض"),
-        "Diagnosis": MessageLookupByLibrary.simpleMessage("التشخيص"),
-        "Disease": MessageLookupByLibrary.simpleMessage("الامراض"),
-        "Dosageform": MessageLookupByLibrary.simpleMessage("طريقه اخذ الجرعه"),
-        "DoyouwanttochoosArteriesandbloodveins":
-            MessageLookupByLibrary.simpleMessage(
-                "هل تريد اختيار الشرايين والأوردة الدموية؟"),
-        "DoyouwanttochooseBones":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار العظام؟"),
-        "DoyouwanttochooseBreast":
-            MessageLookupByLibrary.simpleMessage("هل تريدين اختيار الثدي؟"),
-        "DoyouwanttochooseFemalereproductivesystem":
-            MessageLookupByLibrary.simpleMessage(
-                "هل تريد اختيار الجهاز التناسلي الانثوي؟"),
-        "DoyouwanttochooseKidney":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار الكلى؟"),
-        "DoyouwanttochooseLarynx":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار الحجنجره"),
-        "DoyouwanttochooseMalereproductivesystem":
-            MessageLookupByLibrary.simpleMessage(
-                "هل تريد اختيارالجهاز التناسلي الذكري؟"),
-        "Doyouwanttochooseeyes":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار العين؟"),
-        "DoyouwanttochoosetheBrain":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار المخ؟"),
-        "DoyouwanttochoosetheHeart":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختبار القلب؟"),
-        "DoyouwanttochoosetheLargeandsmallintestine":
-            MessageLookupByLibrary.simpleMessage(
-                "هل تريد اختيار الامعاء الدقيقه و الغليظه؟"),
-        "DoyouwanttochoosetheLiver":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار الكبد؟"),
-        "DoyouwanttochoosetheLungs":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار الرئتين؟"),
-        "DoyouwanttochoosetheNose":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار الأنف؟"),
-        "DoyouwanttochoosetheSpleen":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار الطحال؟"),
-        "DoyouwanttochoosetheStomach":
-            MessageLookupByLibrary.simpleMessage("هل تريد اختيار المعده؟"),
-        "EYE": MessageLookupByLibrary.simpleMessage("العين"),
-        "Egp": MessageLookupByLibrary.simpleMessage("جنيه"),
-        "Femalereproductivesystem":
-            MessageLookupByLibrary.simpleMessage("الجهاز التناسلي الانثوي"),
-        "Find_a_lot_of_specialist_doctors_in_one_place":
-            MessageLookupByLibrary.simpleMessage(
-                "ابحث عن الكثير من الأطباء المتخصصين في مكان واحد"),
-        "FirstAid": MessageLookupByLibrary.simpleMessage("الإسعافات الأولية"),
-        "Get_connect_our_Online_Consultation":
-            MessageLookupByLibrary.simpleMessage(
-                "احصل على اتصال بمشاورتنا عبر الإنترنت"),
-        "Heart": MessageLookupByLibrary.simpleMessage("القلب"),
-        "IsAvaliableInEgypt":
-            MessageLookupByLibrary.simpleMessage("متاح في مصر"),
-        "IsAvaliableInSaudiArabia":
-            MessageLookupByLibrary.simpleMessage("متاح في السعوديه"),
-        "Kidney": MessageLookupByLibrary.simpleMessage("الكلي"),
-        "LabTest": MessageLookupByLibrary.simpleMessage("اختبار\n معملي"),
-        "LabTest2": MessageLookupByLibrary.simpleMessage("اختبار معملي"),
-        "Larynx": MessageLookupByLibrary.simpleMessage("الحنجره"),
-        "Liver": MessageLookupByLibrary.simpleMessage("الكبد"),
-        "Lungs": MessageLookupByLibrary.simpleMessage("الرئه"),
-        "Malereproductivesystem":
-            MessageLookupByLibrary.simpleMessage("الجهاز التناسلي الذكري"),
-        "Medcine": MessageLookupByLibrary.simpleMessage("الادويه\n"),
-        "Medcine1": MessageLookupByLibrary.simpleMessage("الادويه"),
-        "Name1": MessageLookupByLibrary.simpleMessage("الاسم"),
-        "Next": MessageLookupByLibrary.simpleMessage("التالي"),
-        "No": MessageLookupByLibrary.simpleMessage("لا"),
-        "Nose": MessageLookupByLibrary.simpleMessage("الانف"),
-        "Parcode": MessageLookupByLibrary.simpleMessage("باركود"),
-        "Possiblesideeffects":
-            MessageLookupByLibrary.simpleMessage("الآثار الجانبية المحتملة:"),
-        "Possiblesideeffects2":
-            MessageLookupByLibrary.simpleMessage("الآثار الجانبية المحتملة"),
-        "PreventiveMeasures":
-            MessageLookupByLibrary.simpleMessage("الاجراءات الوقائية"),
-        "Price": MessageLookupByLibrary.simpleMessage("السعر:"),
-        "ProductID": MessageLookupByLibrary.simpleMessage("رقم المنتج"),
-        "Productconsistency":
-            MessageLookupByLibrary.simpleMessage("الشكل الدوائي"),
-        "RouteofAdministration":
-            MessageLookupByLibrary.simpleMessage("طريقه أخذ الدواء"),
-        "SearchResults": MessageLookupByLibrary.simpleMessage("نتائج البحث"),
-        "Searchdoctor": MessageLookupByLibrary.simpleMessage("ابحث عن دكتور"),
-        "Searchdrugs":
-            MessageLookupByLibrary.simpleMessage("ابحث عن الأدوية او الفئة..."),
-        "SeeotherCategories":
-            MessageLookupByLibrary.simpleMessage("انظر الي تخصصات أخري"),
-        "SelfCheck": MessageLookupByLibrary.simpleMessage("الفحص الذاتي"),
-        "Showless": MessageLookupByLibrary.simpleMessage("إظهار الاقل"),
-        "Showmore": MessageLookupByLibrary.simpleMessage("إظهار المزيد"),
-        "Skip": MessageLookupByLibrary.simpleMessage("تخطي"),
-        "Smallandlargeintestine":
-            MessageLookupByLibrary.simpleMessage("الامعاء الدقيقه و الغليظه"),
-        "Specifications": MessageLookupByLibrary.simpleMessage("المواصفات"),
-        "Spleen": MessageLookupByLibrary.simpleMessage("الطحال"),
-        "Stomach": MessageLookupByLibrary.simpleMessage("المعده"),
-        "Thelistisempty":
-            MessageLookupByLibrary.simpleMessage("لا يوجد مرض بهذا الاعراض"),
-        "Touchinganorgantoexamineit": MessageLookupByLibrary.simpleMessage(
-            "برجاء لمس العضو الذي تريد فحصه رجاء."),
-        "age": MessageLookupByLibrary.simpleMessage("العمر"),
-        "appointment": MessageLookupByLibrary.simpleMessage("ميعاد"),
-        "canceled": MessageLookupByLibrary.simpleMessage("حجز تم الغاءه"),
-        "change": MessageLookupByLibrary.simpleMessage("تغيير"),
-        "chat": MessageLookupByLibrary.simpleMessage("الدردشات"),
-        "check_yourself": MessageLookupByLibrary.simpleMessage("أكشف\nعن نفسك"),
-        "close": MessageLookupByLibrary.simpleMessage("اغلاق"),
-        "completed": MessageLookupByLibrary.simpleMessage("حجز مكتمل"),
-        "consultation": MessageLookupByLibrary.simpleMessage("استشاره"),
-        "date": MessageLookupByLibrary.simpleMessage("التاريخ"),
-        "do_not_have_an_account":
-            MessageLookupByLibrary.simpleMessage("ليس لديك حساب؟ "),
-        "doctor_detail":
-            MessageLookupByLibrary.simpleMessage("تفاصيل عن الطبيب"),
-        "early_protection_for_your_family_health":
-            MessageLookupByLibrary.simpleMessage(
-                "الحماية المبكرة\n لصحة عائلتك "),
-        "edit_profile":
-            MessageLookupByLibrary.simpleMessage("تعديل الملف الشخصي"),
-        "email_must_not_empty": MessageLookupByLibrary.simpleMessage(
-            "يجب ألا يكون البريد الإلكتروني فارغًا"),
-        "enter_a_valid_email":
-            MessageLookupByLibrary.simpleMessage("أدخل بريد إلكتروني متاح"),
-        "enter_a_valid_password": MessageLookupByLibrary.simpleMessage(
-            "أُووبس! كلمة المرور الخاصة بك غير صحيحة"),
-        "enter_email":
-            MessageLookupByLibrary.simpleMessage("ادخل بريدك الالكتروني"),
-        "enter_name": MessageLookupByLibrary.simpleMessage("ادخل اسمك"),
-        "enter_password":
-            MessageLookupByLibrary.simpleMessage("ادخل كلمة المرور"),
-        "enter_phone": MessageLookupByLibrary.simpleMessage("ادخل رقم هاتفك"),
-        "find_doctor": MessageLookupByLibrary.simpleMessage("ابحث\nعن طبيب"),
-        "find_doctors": MessageLookupByLibrary.simpleMessage("ابحث عن اطباء"),
-        "find_your_desire_health_solution":
-            MessageLookupByLibrary.simpleMessage(
-                "ابحث عن رغبتك\nفي الحل الصحي"),
-        "forgot_password":
-            MessageLookupByLibrary.simpleMessage("نسيت كلمة السر؟"),
-        "go_to_home":
-            MessageLookupByLibrary.simpleMessage("اذهب للصفحه الرئيسيه"),
-        "go_to_login_or_sign_up": MessageLookupByLibrary.simpleMessage(
-            "اذهب لتسجيل الدخول او انشاء الحساب"),
-        "height": MessageLookupByLibrary.simpleMessage("الطول"),
-        "hi": MessageLookupByLibrary.simpleMessage("اهلا"),
-        "language": MessageLookupByLibrary.simpleMessage("اللغه"),
-        "login": MessageLookupByLibrary.simpleMessage("تسجيل الدخول"),
-        "logout": MessageLookupByLibrary.simpleMessage("تسجيل الخروج"),
-        "logoutGoogle":
-            MessageLookupByLibrary.simpleMessage("تسجيل الخروج من جوجل"),
-        "message": MessageLookupByLibrary.simpleMessage("الرسائل"),
-        "messaging": MessageLookupByLibrary.simpleMessage("مراسلة"),
-        "name_must_not_empty":
-            MessageLookupByLibrary.simpleMessage("يجب ألا يكون الاسم فارغًا"),
-        "once_again": MessageLookupByLibrary.simpleMessage(
-            "مره اخري سجلت دخولك بنجاح في تطبيقناالطبي"),
-        "password_must_not_empty": MessageLookupByLibrary.simpleMessage(
-            "يجب ألا تكون كلمة المرور فارغة"),
-        "payment_detail": MessageLookupByLibrary.simpleMessage("تفاصيل الدفع"),
-        "payment_method": MessageLookupByLibrary.simpleMessage("طريقه الدفع"),
-        "please_check_your_Email": MessageLookupByLibrary.simpleMessage(
-            "برجاء التحقق من البريد الالكتروني الخاص بك"),
-        "please_entre": MessageLookupByLibrary.simpleMessage(
-            "برجاء ادخال البريد الالكتروني\nمثل الGmail او Yahoo او Hotmail او ما الي ذلك من بريد الالكتروني"),
-        "producttype": MessageLookupByLibrary.simpleMessage("نوع الدواء"),
-        "profile": MessageLookupByLibrary.simpleMessage("الملف الشخصي"),
-        "reason": MessageLookupByLibrary.simpleMessage("الاسباب"),
-        "recommendedtreatment": MessageLookupByLibrary.simpleMessage("العلاج"),
-        "save": MessageLookupByLibrary.simpleMessage("حفظ"),
-        "schedule": MessageLookupByLibrary.simpleMessage("الحجوزات الخاص بي"),
-        "search_your_doctor":
-            MessageLookupByLibrary.simpleMessage("ابحث عن طبيبك"),
-        "see_all": MessageLookupByLibrary.simpleMessage("اظهار الكل"),
-        "send_code_to_email":
-            MessageLookupByLibrary.simpleMessage("ارسال كود للحساب"),
-        "sign_up": MessageLookupByLibrary.simpleMessage("انشاء حساب"),
-        "something_wrong_happen_please_Login_again":
-            MessageLookupByLibrary.simpleMessage(
-                "شي ما حدث بشكل خاطئ برجاء تسجيل الدخول مره اخري"),
-        "something_wrong_happen_please_sign_up_again":
-            MessageLookupByLibrary.simpleMessage(
-                "حدث خطأ ما، يرجى الاشتراك مرة أخرى"),
-        "symptomsof": MessageLookupByLibrary.simpleMessage("الاعراض"),
-        "top_doctor":
-            MessageLookupByLibrary.simpleMessage("الاطباء الاعلي تقييما"),
-        "total": MessageLookupByLibrary.simpleMessage("اجمالي الحساب"),
-        "upcoming": MessageLookupByLibrary.simpleMessage("الحجوزات القادمه"),
-        "we_have_send_link_to_account_email":
-            MessageLookupByLibrary.simpleMessage(
-                "قمنا بارسال رابط للبريد الالكتروني الخاص بك"),
-        "weight": MessageLookupByLibrary.simpleMessage("الوزن"),
-        "welcome_back": MessageLookupByLibrary.simpleMessage("اهلا بعودتك"),
-        "whenToConsultProfessional":
-            MessageLookupByLibrary.simpleMessage("متي يجب استشارة الطبيب"),
-        "yes": MessageLookupByLibrary.simpleMessage("نعم"),
-        "you_have_account": MessageLookupByLibrary.simpleMessage("لديك حساب؟ "),
-        "you_have_already_signup_successfully_into_medidoc_app":
-            MessageLookupByLibrary.simpleMessage(
-                "لقد قمت بالفعل بالتسجيل بنجا في تطبيقنا")
-      };
+    "ActiveIngredients": MessageLookupByLibrary.simpleMessage("المواد الفعاله"),
+    "AnatomicalLocation": MessageLookupByLibrary.simpleMessage(
+      "الموقع التشريحي",
+    ),
+    "Arteriesandbloodveins": MessageLookupByLibrary.simpleMessage(
+      "الشرايين والأوردة الدموية",
+    ),
+    "Bones": MessageLookupByLibrary.simpleMessage("العظام"),
+    "Brain": MessageLookupByLibrary.simpleMessage("المخ"),
+    "Breast": MessageLookupByLibrary.simpleMessage("صدر"),
+    "Cancel": MessageLookupByLibrary.simpleMessage("الغاء"),
+    "Categories": MessageLookupByLibrary.simpleMessage("التخصصات"),
+    "Causes": MessageLookupByLibrary.simpleMessage("الاسباب"),
+    "Choose": MessageLookupByLibrary.simpleMessage("اختار"),
+    "Class": MessageLookupByLibrary.simpleMessage("العائله"),
+    "Classification": MessageLookupByLibrary.simpleMessage("التصنيف"),
+    "Company": MessageLookupByLibrary.simpleMessage("الشركة"),
+    "Consult_only_with_a_doctor_you_trust":
+        MessageLookupByLibrary.simpleMessage("استشر فقط الطبيب الذي تثق به"),
+    "Description": MessageLookupByLibrary.simpleMessage("الوصف"),
+    "Details": MessageLookupByLibrary.simpleMessage("التفاصيل"),
+    "DetailsAboutDiseases": MessageLookupByLibrary.simpleMessage(
+      "تفاصيل عن الأمراض",
+    ),
+    "Diagnosis": MessageLookupByLibrary.simpleMessage("التشخيص"),
+    "Disease": MessageLookupByLibrary.simpleMessage("الامراض"),
+    "Dosageform": MessageLookupByLibrary.simpleMessage("طريقه اخذ الجرعه"),
+    "DoyouwanttochoosArteriesandbloodveins":
+        MessageLookupByLibrary.simpleMessage(
+          "هل تريد اختيار الشرايين والأوردة الدموية؟",
+        ),
+    "DoyouwanttochooseBones": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار العظام؟",
+    ),
+    "DoyouwanttochooseBreast": MessageLookupByLibrary.simpleMessage(
+      "هل تريدين اختيار الثدي؟",
+    ),
+    "DoyouwanttochooseFemalereproductivesystem":
+        MessageLookupByLibrary.simpleMessage(
+          "هل تريد اختيار الجهاز التناسلي الانثوي؟",
+        ),
+    "DoyouwanttochooseKidney": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار الكلى؟",
+    ),
+    "DoyouwanttochooseLarynx": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار الحجنجره",
+    ),
+    "DoyouwanttochooseMalereproductivesystem":
+        MessageLookupByLibrary.simpleMessage(
+          "هل تريد اختيارالجهاز التناسلي الذكري؟",
+        ),
+    "Doyouwanttochooseeyes": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار العين؟",
+    ),
+    "DoyouwanttochoosetheBrain": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار المخ؟",
+    ),
+    "DoyouwanttochoosetheHeart": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختبار القلب؟",
+    ),
+    "DoyouwanttochoosetheLargeandsmallintestine":
+        MessageLookupByLibrary.simpleMessage(
+          "هل تريد اختيار الامعاء الدقيقه و الغليظه؟",
+        ),
+    "DoyouwanttochoosetheLiver": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار الكبد؟",
+    ),
+    "DoyouwanttochoosetheLungs": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار الرئتين؟",
+    ),
+    "DoyouwanttochoosetheNose": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار الأنف؟",
+    ),
+    "DoyouwanttochoosetheSpleen": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار الطحال؟",
+    ),
+    "DoyouwanttochoosetheStomach": MessageLookupByLibrary.simpleMessage(
+      "هل تريد اختيار المعده؟",
+    ),
+    "EYE": MessageLookupByLibrary.simpleMessage("العين"),
+    "Egp": MessageLookupByLibrary.simpleMessage("جنيه"),
+    "Femalereproductivesystem": MessageLookupByLibrary.simpleMessage(
+      "الجهاز التناسلي الانثوي",
+    ),
+    "Find_a_lot_of_specialist_doctors_in_one_place":
+        MessageLookupByLibrary.simpleMessage(
+          "ابحث عن الكثير من الأطباء المتخصصين في مكان واحد",
+        ),
+    "FirstAid": MessageLookupByLibrary.simpleMessage("الإسعافات الأولية"),
+    "Get_connect_our_Online_Consultation": MessageLookupByLibrary.simpleMessage(
+      "احصل على اتصال بمشاورتنا عبر الإنترنت",
+    ),
+    "Heart": MessageLookupByLibrary.simpleMessage("القلب"),
+    "IsAvaliableInEgypt": MessageLookupByLibrary.simpleMessage("متاح في مصر"),
+    "IsAvaliableInSaudiArabia": MessageLookupByLibrary.simpleMessage(
+      "متاح في السعوديه",
+    ),
+    "Kidney": MessageLookupByLibrary.simpleMessage("الكلي"),
+    "LabTest": MessageLookupByLibrary.simpleMessage("اختبار\n معملي"),
+    "LabTest2": MessageLookupByLibrary.simpleMessage("اختبار معملي"),
+    "Larynx": MessageLookupByLibrary.simpleMessage("الحنجره"),
+    "Liver": MessageLookupByLibrary.simpleMessage("الكبد"),
+    "Lungs": MessageLookupByLibrary.simpleMessage("الرئه"),
+    "Malereproductivesystem": MessageLookupByLibrary.simpleMessage(
+      "الجهاز التناسلي الذكري",
+    ),
+    "Medcine": MessageLookupByLibrary.simpleMessage("الادويه\n"),
+    "Medcine1": MessageLookupByLibrary.simpleMessage("الادويه"),
+    "Name1": MessageLookupByLibrary.simpleMessage("الاسم"),
+    "Next": MessageLookupByLibrary.simpleMessage("التالي"),
+    "No": MessageLookupByLibrary.simpleMessage("لا"),
+    "Nose": MessageLookupByLibrary.simpleMessage("الانف"),
+    "Parcode": MessageLookupByLibrary.simpleMessage("باركود"),
+    "Possiblesideeffects": MessageLookupByLibrary.simpleMessage(
+      "الآثار الجانبية المحتملة:",
+    ),
+    "Possiblesideeffects2": MessageLookupByLibrary.simpleMessage(
+      "الآثار الجانبية المحتملة",
+    ),
+    "PreventiveMeasures": MessageLookupByLibrary.simpleMessage(
+      "الاجراءات الوقائية",
+    ),
+    "Price": MessageLookupByLibrary.simpleMessage("السعر:"),
+    "ProductID": MessageLookupByLibrary.simpleMessage("رقم المنتج"),
+    "Productconsistency": MessageLookupByLibrary.simpleMessage("الشكل الدوائي"),
+    "RouteofAdministration": MessageLookupByLibrary.simpleMessage(
+      "طريقه أخذ الدواء",
+    ),
+    "SearchResults": MessageLookupByLibrary.simpleMessage("نتائج البحث"),
+    "Searchdoctor": MessageLookupByLibrary.simpleMessage("ابحث عن دكتور"),
+    "Searchdrugs": MessageLookupByLibrary.simpleMessage(
+      "ابحث عن الأدوية او الفئة...",
+    ),
+    "SeeotherCategories": MessageLookupByLibrary.simpleMessage(
+      "انظر الي تخصصات أخري",
+    ),
+    "SelfCheck": MessageLookupByLibrary.simpleMessage("الفحص الذاتي"),
+    "Showless": MessageLookupByLibrary.simpleMessage("إظهار الاقل"),
+    "Showmore": MessageLookupByLibrary.simpleMessage("إظهار المزيد"),
+    "Skip": MessageLookupByLibrary.simpleMessage("تخطي"),
+    "Smallandlargeintestine": MessageLookupByLibrary.simpleMessage(
+      "الامعاء الدقيقه و الغليظه",
+    ),
+    "Specifications": MessageLookupByLibrary.simpleMessage("المواصفات"),
+    "Spleen": MessageLookupByLibrary.simpleMessage("الطحال"),
+    "Stomach": MessageLookupByLibrary.simpleMessage("المعده"),
+    "Thelistisempty": MessageLookupByLibrary.simpleMessage(
+      "لا يوجد مرض بهذا الاعراض",
+    ),
+    "Touchinganorgantoexamineit": MessageLookupByLibrary.simpleMessage(
+      "برجاء لمس العضو الذي تريد فحصه رجاء.",
+    ),
+    "age": MessageLookupByLibrary.simpleMessage("العمر"),
+    "appointment": MessageLookupByLibrary.simpleMessage("ميعاد"),
+    "canceled": MessageLookupByLibrary.simpleMessage("حجز تم الغاءه"),
+    "change": MessageLookupByLibrary.simpleMessage("تغيير"),
+    "chat": MessageLookupByLibrary.simpleMessage("الدردشات"),
+    "check_yourself": MessageLookupByLibrary.simpleMessage("أكشف\nعن نفسك"),
+    "close": MessageLookupByLibrary.simpleMessage("اغلاق"),
+    "completed": MessageLookupByLibrary.simpleMessage("حجز مكتمل"),
+    "consultation": MessageLookupByLibrary.simpleMessage("استشاره"),
+    "date": MessageLookupByLibrary.simpleMessage("التاريخ"),
+    "do_not_have_an_account": MessageLookupByLibrary.simpleMessage(
+      "ليس لديك حساب؟ ",
+    ),
+    "doctor_detail": MessageLookupByLibrary.simpleMessage("تفاصيل عن الطبيب"),
+    "early_protection_for_your_family_health":
+        MessageLookupByLibrary.simpleMessage("الحماية المبكرة\n لصحة عائلتك "),
+    "edit_profile": MessageLookupByLibrary.simpleMessage("تعديل الملف الشخصي"),
+    "email_must_not_empty": MessageLookupByLibrary.simpleMessage(
+      "يجب ألا يكون البريد الإلكتروني فارغًا",
+    ),
+    "enter_a_valid_email": MessageLookupByLibrary.simpleMessage(
+      "أدخل بريد إلكتروني متاح",
+    ),
+    "enter_a_valid_password": MessageLookupByLibrary.simpleMessage(
+      "أُووبس! كلمة المرور الخاصة بك غير صحيحة",
+    ),
+    "enter_email": MessageLookupByLibrary.simpleMessage(
+      "ادخل بريدك الالكتروني",
+    ),
+    "enter_name": MessageLookupByLibrary.simpleMessage("ادخل اسمك"),
+    "enter_password": MessageLookupByLibrary.simpleMessage("ادخل كلمة المرور"),
+    "enter_phone": MessageLookupByLibrary.simpleMessage("ادخل رقم هاتفك"),
+    "find_doctor": MessageLookupByLibrary.simpleMessage("ابحث\nعن طبيب"),
+    "find_doctors": MessageLookupByLibrary.simpleMessage("ابحث عن اطباء"),
+    "find_your_desire_health_solution": MessageLookupByLibrary.simpleMessage(
+      "ابحث عن رغبتك\nفي الحل الصحي",
+    ),
+    "forgot_password": MessageLookupByLibrary.simpleMessage("نسيت كلمة السر؟"),
+    "go_to_home": MessageLookupByLibrary.simpleMessage("اذهب للصفحه الرئيسيه"),
+    "go_to_login_or_sign_up": MessageLookupByLibrary.simpleMessage(
+      "اذهب لتسجيل الدخول او انشاء الحساب",
+    ),
+    "height": MessageLookupByLibrary.simpleMessage("الطول"),
+    "hi": MessageLookupByLibrary.simpleMessage("اهلا"),
+    "language": MessageLookupByLibrary.simpleMessage("اللغه"),
+    "login": MessageLookupByLibrary.simpleMessage("تسجيل الدخول"),
+    "logout": MessageLookupByLibrary.simpleMessage("تسجيل الخروج"),
+    "logoutGoogle": MessageLookupByLibrary.simpleMessage(
+      "تسجيل الخروج من جوجل",
+    ),
+    "message": MessageLookupByLibrary.simpleMessage("الرسائل"),
+    "messaging": MessageLookupByLibrary.simpleMessage("مراسلة"),
+    "name_must_not_empty": MessageLookupByLibrary.simpleMessage(
+      "يجب ألا يكون الاسم فارغًا",
+    ),
+    "nurse": MessageLookupByLibrary.simpleMessage("الممرضين"),
+    "once_again": MessageLookupByLibrary.simpleMessage(
+      "مره اخري سجلت دخولك بنجاح في تطبيقناالطبي",
+    ),
+    "password_must_not_empty": MessageLookupByLibrary.simpleMessage(
+      "يجب ألا تكون كلمة المرور فارغة",
+    ),
+    "payment_detail": MessageLookupByLibrary.simpleMessage("تفاصيل الدفع"),
+    "payment_method": MessageLookupByLibrary.simpleMessage("طريقه الدفع"),
+    "please_check_your_Email": MessageLookupByLibrary.simpleMessage(
+      "برجاء التحقق من البريد الالكتروني الخاص بك",
+    ),
+    "please_entre": MessageLookupByLibrary.simpleMessage(
+      "برجاء ادخال البريد الالكتروني\nمثل الGmail او Yahoo او Hotmail او ما الي ذلك من بريد الالكتروني",
+    ),
+    "producttype": MessageLookupByLibrary.simpleMessage("نوع الدواء"),
+    "profile": MessageLookupByLibrary.simpleMessage("الملف الشخصي"),
+    "reason": MessageLookupByLibrary.simpleMessage("الاسباب"),
+    "recommendedtreatment": MessageLookupByLibrary.simpleMessage("العلاج"),
+    "save": MessageLookupByLibrary.simpleMessage("حفظ"),
+    "schedule": MessageLookupByLibrary.simpleMessage("الحجوزات الخاص بي"),
+    "search_your_doctor": MessageLookupByLibrary.simpleMessage("ابحث عن طبيبك"),
+    "see_all": MessageLookupByLibrary.simpleMessage("اظهار الكل"),
+    "send_code_to_email": MessageLookupByLibrary.simpleMessage(
+      "ارسال كود للحساب",
+    ),
+    "sign_up": MessageLookupByLibrary.simpleMessage("انشاء حساب"),
+    "something_wrong_happen_please_Login_again":
+        MessageLookupByLibrary.simpleMessage(
+          "شي ما حدث بشكل خاطئ برجاء تسجيل الدخول مره اخري",
+        ),
+    "something_wrong_happen_please_sign_up_again":
+        MessageLookupByLibrary.simpleMessage(
+          "حدث خطأ ما، يرجى الاشتراك مرة أخرى",
+        ),
+    "symptomsof": MessageLookupByLibrary.simpleMessage("الاعراض"),
+    "top_doctor": MessageLookupByLibrary.simpleMessage("الاطباء الاعلي تقييما"),
+    "total": MessageLookupByLibrary.simpleMessage("اجمالي الحساب"),
+    "upcoming": MessageLookupByLibrary.simpleMessage("الحجوزات القادمه"),
+    "we_have_send_link_to_account_email": MessageLookupByLibrary.simpleMessage(
+      "قمنا بارسال رابط للبريد الالكتروني الخاص بك",
+    ),
+    "weight": MessageLookupByLibrary.simpleMessage("الوزن"),
+    "welcome_back": MessageLookupByLibrary.simpleMessage("اهلا بعودتك"),
+    "whenToConsultProfessional": MessageLookupByLibrary.simpleMessage(
+      "متي يجب استشارة الطبيب",
+    ),
+    "yes": MessageLookupByLibrary.simpleMessage("نعم"),
+    "you_have_account": MessageLookupByLibrary.simpleMessage("لديك حساب؟ "),
+    "you_have_already_signup_successfully_into_medidoc_app":
+        MessageLookupByLibrary.simpleMessage(
+          "لقد قمت بالفعل بالتسجيل بنجا في تطبيقنا",
+        ),
+  };
 }
