@@ -22,232 +22,282 @@ class MessageLookup extends MessageLookupByLibrary {
 
   final messages = _notInlinedMessages(_notInlinedMessages);
   static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
-        "ActiveIngredients":
-            MessageLookupByLibrary.simpleMessage("Active Ingredients"),
-        "AnatomicalLocation":
-            MessageLookupByLibrary.simpleMessage("AnatomicalLocation"),
-        "Arteriesandbloodveins":
-            MessageLookupByLibrary.simpleMessage("Arteries and blood veins"),
-        "Bones": MessageLookupByLibrary.simpleMessage("Bones"),
-        "Brain": MessageLookupByLibrary.simpleMessage("Brain"),
-        "Breast": MessageLookupByLibrary.simpleMessage("Breast"),
-        "Cancel": MessageLookupByLibrary.simpleMessage("Cancel"),
-        "Categories": MessageLookupByLibrary.simpleMessage("Categories"),
-        "Causes": MessageLookupByLibrary.simpleMessage("Causes"),
-        "Choose": MessageLookupByLibrary.simpleMessage("Choose"),
-        "Class": MessageLookupByLibrary.simpleMessage("Class"),
-        "Classification":
-            MessageLookupByLibrary.simpleMessage("Classification"),
-        "Company": MessageLookupByLibrary.simpleMessage("Company"),
-        "Consult_only_with_a_doctor_you_trust":
-            MessageLookupByLibrary.simpleMessage(
-                "Consult only with a doctor you trust"),
-        "Description": MessageLookupByLibrary.simpleMessage("Description"),
-        "Details": MessageLookupByLibrary.simpleMessage("Details"),
-        "DetailsAboutDiseases":
-            MessageLookupByLibrary.simpleMessage("Details About Diseases"),
-        "Diagnosis": MessageLookupByLibrary.simpleMessage("Diagnosis"),
-        "Disease": MessageLookupByLibrary.simpleMessage("Disease"),
-        "Dosageform": MessageLookupByLibrary.simpleMessage("Dosage form"),
-        "DoyouwanttochoosArteriesandbloodveins":
-            MessageLookupByLibrary.simpleMessage(
-                "Do you want to choose Arteries and blood veins?"),
-        "DoyouwanttochooseBones": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose Bones?"),
-        "DoyouwanttochooseBreast": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose Breast?"),
-        "DoyouwanttochooseFemalereproductivesystem":
-            MessageLookupByLibrary.simpleMessage(
-                "Do you want to choose Female reproductive system?"),
-        "DoyouwanttochooseKidney": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose Kidney?"),
-        "DoyouwanttochooseLarynx": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose Larynx?"),
-        "DoyouwanttochooseMalereproductivesystem":
-            MessageLookupByLibrary.simpleMessage(
-                "Do you want to choose Male reproductive system?"),
-        "Doyouwanttochooseeyes":
-            MessageLookupByLibrary.simpleMessage("Do you want to choose eyes?"),
-        "DoyouwanttochoosetheBrain": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose theBrain?"),
-        "DoyouwanttochoosetheHeart": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose the Heart?"),
-        "DoyouwanttochoosetheLargeandsmallintestine":
-            MessageLookupByLibrary.simpleMessage(
-                "Do you want to choose the Large and small intestine?"),
-        "DoyouwanttochoosetheLiver": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose the Liver?"),
-        "DoyouwanttochoosetheLungs": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose the Lungs?"),
-        "DoyouwanttochoosetheNose": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose the Nose?"),
-        "DoyouwanttochoosetheSpleen": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose the Spleen?"),
-        "DoyouwanttochoosetheStomach": MessageLookupByLibrary.simpleMessage(
-            "Do you want to choose the Stomach?"),
-        "EYE": MessageLookupByLibrary.simpleMessage("EYE"),
-        "Egp": MessageLookupByLibrary.simpleMessage("Egp"),
-        "Femalereproductivesystem":
-            MessageLookupByLibrary.simpleMessage("Female Reproductive System"),
-        "Find_a_lot_of_specialist_doctors_in_one_place":
-            MessageLookupByLibrary.simpleMessage(
-                "Find a lot of specialist doctors in one place"),
-        "FirstAid": MessageLookupByLibrary.simpleMessage("First Aid"),
-        "Get_connect_our_Online_Consultation":
-            MessageLookupByLibrary.simpleMessage(
-                "Get connect our Online Consultation"),
-        "Heart": MessageLookupByLibrary.simpleMessage("Heart"),
-        "IsAvaliableInEgypt":
-            MessageLookupByLibrary.simpleMessage("Is Avaliable In Egypt"),
-        "IsAvaliableInSaudiArabia":
-            MessageLookupByLibrary.simpleMessage("Is Avaliable In SaudiArabia"),
-        "Kidney": MessageLookupByLibrary.simpleMessage("Kidney"),
-        "LabTest": MessageLookupByLibrary.simpleMessage("Lab\nTest"),
-        "LabTest2": MessageLookupByLibrary.simpleMessage("Lab Test"),
-        "Larynx": MessageLookupByLibrary.simpleMessage("Throat"),
-        "Liver": MessageLookupByLibrary.simpleMessage("Liver"),
-        "Lungs": MessageLookupByLibrary.simpleMessage("Lungs"),
-        "Malereproductivesystem":
-            MessageLookupByLibrary.simpleMessage("Male Reproductive System"),
-        "Medcine": MessageLookupByLibrary.simpleMessage("Medcine\n "),
-        "Medcine1": MessageLookupByLibrary.simpleMessage("Medcine"),
-        "Name1": MessageLookupByLibrary.simpleMessage("Name"),
-        "Next": MessageLookupByLibrary.simpleMessage("Next"),
-        "No": MessageLookupByLibrary.simpleMessage("No"),
-        "Nose": MessageLookupByLibrary.simpleMessage("Nose"),
-        "Parcode": MessageLookupByLibrary.simpleMessage("Parcode"),
-        "Possiblesideeffects":
-            MessageLookupByLibrary.simpleMessage("Possible side effects:"),
-        "Possiblesideeffects2":
-            MessageLookupByLibrary.simpleMessage("Possible side effects"),
-        "PreventiveMeasures":
-            MessageLookupByLibrary.simpleMessage("Preventive Measures"),
-        "Price": MessageLookupByLibrary.simpleMessage("Price:"),
-        "ProductID": MessageLookupByLibrary.simpleMessage("Product ID"),
-        "Productconsistency":
-            MessageLookupByLibrary.simpleMessage("Product Consistency"),
-        "RouteofAdministration":
-            MessageLookupByLibrary.simpleMessage("Route of Administration"),
-        "SearchResults": MessageLookupByLibrary.simpleMessage("Search Results"),
-        "Searchdoctor": MessageLookupByLibrary.simpleMessage("Find a doctor"),
-        "Searchdrugs":
-            MessageLookupByLibrary.simpleMessage("Search drugs, category..."),
-        "SeeotherCategories":
-            MessageLookupByLibrary.simpleMessage("See other Categories"),
-        "SelfCheck": MessageLookupByLibrary.simpleMessage("Self-Check"),
-        "Showless": MessageLookupByLibrary.simpleMessage("Show less"),
-        "Showmore": MessageLookupByLibrary.simpleMessage("Show more"),
-        "Skip": MessageLookupByLibrary.simpleMessage("Skip"),
-        "Smallandlargeintestine":
-            MessageLookupByLibrary.simpleMessage("Small and large intestine"),
-        "Specifications":
-            MessageLookupByLibrary.simpleMessage("Specifications"),
-        "Spleen": MessageLookupByLibrary.simpleMessage("Spleen"),
-        "Stomach": MessageLookupByLibrary.simpleMessage("Stomach"),
-        "Thelistisempty": MessageLookupByLibrary.simpleMessage(
-            "There is no disease with these symptoms"),
-        "Touchinganorgantoexamineit": MessageLookupByLibrary.simpleMessage(
-            "Touching an organ to examine it."),
-        "age": MessageLookupByLibrary.simpleMessage("Age"),
-        "appointment": MessageLookupByLibrary.simpleMessage("Appointment"),
-        "canceled": MessageLookupByLibrary.simpleMessage("Canceled"),
-        "change": MessageLookupByLibrary.simpleMessage("Change"),
-        "chat": MessageLookupByLibrary.simpleMessage("Chats"),
-        "check_yourself":
-            MessageLookupByLibrary.simpleMessage("Check\nYourself"),
-        "close": MessageLookupByLibrary.simpleMessage("Close"),
-        "completed": MessageLookupByLibrary.simpleMessage("Completed"),
-        "consultation": MessageLookupByLibrary.simpleMessage("Consultation"),
-        "date": MessageLookupByLibrary.simpleMessage("Date"),
-        "do_not_have_an_account":
-            MessageLookupByLibrary.simpleMessage("Don\'t have an account?"),
-        "doctor_detail": MessageLookupByLibrary.simpleMessage("Doctor Detail"),
-        "early_protection_for_your_family_health":
-            MessageLookupByLibrary.simpleMessage(
-                "Early protection for\nyour family health"),
-        "edit_profile": MessageLookupByLibrary.simpleMessage("Edit Profile"),
-        "email_must_not_empty":
-            MessageLookupByLibrary.simpleMessage("Email must not empty"),
-        "enter_a_valid_email":
-            MessageLookupByLibrary.simpleMessage("Enter a valid email"),
-        "enter_a_valid_password": MessageLookupByLibrary.simpleMessage(
-            "Oops! Your Password Is Not Correct"),
-        "enter_email": MessageLookupByLibrary.simpleMessage("Enter your Email"),
-        "enter_name": MessageLookupByLibrary.simpleMessage("Enter your Name"),
-        "enter_password":
-            MessageLookupByLibrary.simpleMessage("Enter your Password"),
-        "enter_phone": MessageLookupByLibrary.simpleMessage("Enter your phone"),
-        "find_doctor": MessageLookupByLibrary.simpleMessage("Find\nDoctor"),
-        "find_doctors": MessageLookupByLibrary.simpleMessage("Find Doctors"),
-        "find_your_desire_health_solution":
-            MessageLookupByLibrary.simpleMessage(
-                "Find your desire \nfor your health solution"),
-        "forgot_password":
-            MessageLookupByLibrary.simpleMessage("Forgot password?"),
-        "go_to_home": MessageLookupByLibrary.simpleMessage("Go to home"),
-        "go_to_login_or_sign_up":
-            MessageLookupByLibrary.simpleMessage("Go to Login or Signup"),
-        "height": MessageLookupByLibrary.simpleMessage("Height"),
-        "hi": MessageLookupByLibrary.simpleMessage("Hi"),
-        "language": MessageLookupByLibrary.simpleMessage("Language"),
-        "login": MessageLookupByLibrary.simpleMessage("Login"),
-        "logout": MessageLookupByLibrary.simpleMessage("Logout"),
-        "logoutGoogle":
-            MessageLookupByLibrary.simpleMessage("Logout from google"),
-        "message": MessageLookupByLibrary.simpleMessage("Message"),
-        "messaging": MessageLookupByLibrary.simpleMessage("Messaging"),
-        "name_must_not_empty":
-            MessageLookupByLibrary.simpleMessage("Name must not empty"),
-        "once_again": MessageLookupByLibrary.simpleMessage(
-            "Once again you login successfully into medical app"),
-        "password_must_not_empty":
-            MessageLookupByLibrary.simpleMessage("Password must not empty"),
-        "payment_detail":
-            MessageLookupByLibrary.simpleMessage("Payment Detail"),
-        "payment_method":
-            MessageLookupByLibrary.simpleMessage("Payment Method"),
-        "phone_must_not_empty":
-            MessageLookupByLibrary.simpleMessage("phone must not empty"),
-        "please_check_your_Email":
-            MessageLookupByLibrary.simpleMessage("Please Check your Email"),
-        "please_entre": MessageLookupByLibrary.simpleMessage(
-            "Please Entre E-mail\nit like gmail,yahoo,Hotmail,etc..."),
-        "producttype": MessageLookupByLibrary.simpleMessage("Product Type"),
-        "profile": MessageLookupByLibrary.simpleMessage("profile"),
-        "reason": MessageLookupByLibrary.simpleMessage("Reason"),
-        "recommendedtreatment":
-            MessageLookupByLibrary.simpleMessage("Recommended Treatment"),
-        "save": MessageLookupByLibrary.simpleMessage("Save"),
-        "schedule": MessageLookupByLibrary.simpleMessage("Schedule"),
-        "search_your_doctor":
-            MessageLookupByLibrary.simpleMessage("Search your doctor"),
-        "see_all": MessageLookupByLibrary.simpleMessage("Show all"),
-        "send_code_to_email":
-            MessageLookupByLibrary.simpleMessage("Send code to Email"),
-        "sign_up": MessageLookupByLibrary.simpleMessage("Sign Up"),
-        "something_wrong_happen_please_Login_again":
-            MessageLookupByLibrary.simpleMessage(
-                "Something wrong happen Please Login again"),
-        "something_wrong_happen_please_sign_up_again":
-            MessageLookupByLibrary.simpleMessage(
-                "Something wrong happen Please Sign up again"),
-        "symptomsof": MessageLookupByLibrary.simpleMessage("symptoms"),
-        "top_doctor": MessageLookupByLibrary.simpleMessage("Top Doctor"),
-        "total": MessageLookupByLibrary.simpleMessage("Total"),
-        "upcoming": MessageLookupByLibrary.simpleMessage("Upcoming"),
-        "we_have_send_link_to_account_email":
-            MessageLookupByLibrary.simpleMessage(
-                "we have send link to account Email"),
-        "weight": MessageLookupByLibrary.simpleMessage("Weight"),
-        "welcome_back":
-            MessageLookupByLibrary.simpleMessage("Yeay! Welcome Back"),
-        "whenToConsultProfessional": MessageLookupByLibrary.simpleMessage(
-            "when To Consult Professional"),
-        "yes": MessageLookupByLibrary.simpleMessage("Yes"),
-        "you_have_account":
-            MessageLookupByLibrary.simpleMessage("You have account? "),
-        "you_have_already_signup_successfully_into_medidoc_app":
-            MessageLookupByLibrary.simpleMessage(
-                "You have already SignUp successfully into medidoc app")
-      };
+    "ActiveIngredients": MessageLookupByLibrary.simpleMessage(
+      "Active Ingredients",
+    ),
+    "AnatomicalLocation": MessageLookupByLibrary.simpleMessage(
+      "AnatomicalLocation",
+    ),
+    "Arteriesandbloodveins": MessageLookupByLibrary.simpleMessage(
+      "Arteries and blood veins",
+    ),
+    "Bones": MessageLookupByLibrary.simpleMessage("Bones"),
+    "Brain": MessageLookupByLibrary.simpleMessage("Brain"),
+    "Breast": MessageLookupByLibrary.simpleMessage("Breast"),
+    "Cancel": MessageLookupByLibrary.simpleMessage("Cancel"),
+    "Categories": MessageLookupByLibrary.simpleMessage("Categories"),
+    "Causes": MessageLookupByLibrary.simpleMessage("Causes"),
+    "Choose": MessageLookupByLibrary.simpleMessage("Choose"),
+    "Class": MessageLookupByLibrary.simpleMessage("Class"),
+    "Classification": MessageLookupByLibrary.simpleMessage("Classification"),
+    "Company": MessageLookupByLibrary.simpleMessage("Company"),
+    "Consult_only_with_a_doctor_you_trust":
+        MessageLookupByLibrary.simpleMessage(
+          "Consult only with a doctor you trust",
+        ),
+    "Description": MessageLookupByLibrary.simpleMessage("Description"),
+    "Details": MessageLookupByLibrary.simpleMessage("Details"),
+    "DetailsAboutDiseases": MessageLookupByLibrary.simpleMessage(
+      "Details About Diseases",
+    ),
+    "Diagnosis": MessageLookupByLibrary.simpleMessage("Diagnosis"),
+    "Disease": MessageLookupByLibrary.simpleMessage("Disease"),
+    "Dosageform": MessageLookupByLibrary.simpleMessage("Dosage form"),
+    "DoyouwanttochoosArteriesandbloodveins":
+        MessageLookupByLibrary.simpleMessage(
+          "Do you want to choose Arteries and blood veins?",
+        ),
+    "DoyouwanttochooseBones": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose Bones?",
+    ),
+    "DoyouwanttochooseBreast": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose Breast?",
+    ),
+    "DoyouwanttochooseFemalereproductivesystem":
+        MessageLookupByLibrary.simpleMessage(
+          "Do you want to choose Female reproductive system?",
+        ),
+    "DoyouwanttochooseKidney": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose Kidney?",
+    ),
+    "DoyouwanttochooseLarynx": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose Larynx?",
+    ),
+    "DoyouwanttochooseMalereproductivesystem":
+        MessageLookupByLibrary.simpleMessage(
+          "Do you want to choose Male reproductive system?",
+        ),
+    "Doyouwanttochooseeyes": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose eyes?",
+    ),
+    "DoyouwanttochoosetheBrain": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose theBrain?",
+    ),
+    "DoyouwanttochoosetheHeart": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose the Heart?",
+    ),
+    "DoyouwanttochoosetheLargeandsmallintestine":
+        MessageLookupByLibrary.simpleMessage(
+          "Do you want to choose the Large and small intestine?",
+        ),
+    "DoyouwanttochoosetheLiver": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose the Liver?",
+    ),
+    "DoyouwanttochoosetheLungs": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose the Lungs?",
+    ),
+    "DoyouwanttochoosetheNose": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose the Nose?",
+    ),
+    "DoyouwanttochoosetheSpleen": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose the Spleen?",
+    ),
+    "DoyouwanttochoosetheStomach": MessageLookupByLibrary.simpleMessage(
+      "Do you want to choose the Stomach?",
+    ),
+    "EYE": MessageLookupByLibrary.simpleMessage("EYE"),
+    "Egp": MessageLookupByLibrary.simpleMessage("Egp"),
+    "Femalereproductivesystem": MessageLookupByLibrary.simpleMessage(
+      "Female Reproductive System",
+    ),
+    "Find_a_lot_of_specialist_doctors_in_one_place":
+        MessageLookupByLibrary.simpleMessage(
+          "Find a lot of specialist doctors in one place",
+        ),
+    "FirstAid": MessageLookupByLibrary.simpleMessage("First Aid"),
+    "Get_connect_our_Online_Consultation": MessageLookupByLibrary.simpleMessage(
+      "Get connect our Online Consultation",
+    ),
+    "Heart": MessageLookupByLibrary.simpleMessage("Heart"),
+    "IsAvaliableInEgypt": MessageLookupByLibrary.simpleMessage(
+      "Is Avaliable In Egypt",
+    ),
+    "IsAvaliableInSaudiArabia": MessageLookupByLibrary.simpleMessage(
+      "Is Avaliable In SaudiArabia",
+    ),
+    "Kidney": MessageLookupByLibrary.simpleMessage("Kidney"),
+    "LabTest": MessageLookupByLibrary.simpleMessage("Lab\nTest"),
+    "LabTest2": MessageLookupByLibrary.simpleMessage("Lab Test"),
+    "Larynx": MessageLookupByLibrary.simpleMessage("Throat"),
+    "Liver": MessageLookupByLibrary.simpleMessage("Liver"),
+    "Lungs": MessageLookupByLibrary.simpleMessage("Lungs"),
+    "Malereproductivesystem": MessageLookupByLibrary.simpleMessage(
+      "Male Reproductive System",
+    ),
+    "Medcine": MessageLookupByLibrary.simpleMessage("Medcine\n "),
+    "Medcine1": MessageLookupByLibrary.simpleMessage("Medcine"),
+    "Name1": MessageLookupByLibrary.simpleMessage("Name"),
+    "Next": MessageLookupByLibrary.simpleMessage("Next"),
+    "No": MessageLookupByLibrary.simpleMessage("No"),
+    "Nose": MessageLookupByLibrary.simpleMessage("Nose"),
+    "Parcode": MessageLookupByLibrary.simpleMessage("Parcode"),
+    "Possiblesideeffects": MessageLookupByLibrary.simpleMessage(
+      "Possible side effects:",
+    ),
+    "Possiblesideeffects2": MessageLookupByLibrary.simpleMessage(
+      "Possible side effects",
+    ),
+    "PreventiveMeasures": MessageLookupByLibrary.simpleMessage(
+      "Preventive Measures",
+    ),
+    "Price": MessageLookupByLibrary.simpleMessage("Price:"),
+    "ProductID": MessageLookupByLibrary.simpleMessage("Product ID"),
+    "Productconsistency": MessageLookupByLibrary.simpleMessage(
+      "Product Consistency",
+    ),
+    "RouteofAdministration": MessageLookupByLibrary.simpleMessage(
+      "Route of Administration",
+    ),
+    "SearchResults": MessageLookupByLibrary.simpleMessage("Search Results"),
+    "Searchdoctor": MessageLookupByLibrary.simpleMessage("Find a doctor"),
+    "Searchdrugs": MessageLookupByLibrary.simpleMessage(
+      "Search drugs, category...",
+    ),
+    "SeeotherCategories": MessageLookupByLibrary.simpleMessage(
+      "See other Categories",
+    ),
+    "SelfCheck": MessageLookupByLibrary.simpleMessage("Self-Check"),
+    "Showless": MessageLookupByLibrary.simpleMessage("Show less"),
+    "Showmore": MessageLookupByLibrary.simpleMessage("Show more"),
+    "Skip": MessageLookupByLibrary.simpleMessage("Skip"),
+    "Smallandlargeintestine": MessageLookupByLibrary.simpleMessage(
+      "Small and large intestine",
+    ),
+    "Specifications": MessageLookupByLibrary.simpleMessage("Specifications"),
+    "Spleen": MessageLookupByLibrary.simpleMessage("Spleen"),
+    "Stomach": MessageLookupByLibrary.simpleMessage("Stomach"),
+    "Thelistisempty": MessageLookupByLibrary.simpleMessage(
+      "There is no disease with these symptoms",
+    ),
+    "Touchinganorgantoexamineit": MessageLookupByLibrary.simpleMessage(
+      "Touching an organ to examine it.",
+    ),
+    "age": MessageLookupByLibrary.simpleMessage("Age"),
+    "appointment": MessageLookupByLibrary.simpleMessage("Appointment"),
+    "canceled": MessageLookupByLibrary.simpleMessage("Canceled"),
+    "change": MessageLookupByLibrary.simpleMessage("Change"),
+    "chat": MessageLookupByLibrary.simpleMessage("Chats"),
+    "check_yourself": MessageLookupByLibrary.simpleMessage("Check\nYourself"),
+    "close": MessageLookupByLibrary.simpleMessage("Close"),
+    "completed": MessageLookupByLibrary.simpleMessage("Completed"),
+    "consultation": MessageLookupByLibrary.simpleMessage("Consultation"),
+    "date": MessageLookupByLibrary.simpleMessage("Date"),
+    "do_not_have_an_account": MessageLookupByLibrary.simpleMessage(
+      "Don\'t have an account?",
+    ),
+    "doctor_detail": MessageLookupByLibrary.simpleMessage("Doctor Detail"),
+    "early_protection_for_your_family_health":
+        MessageLookupByLibrary.simpleMessage(
+          "Early protection for\nyour family health",
+        ),
+    "edit_profile": MessageLookupByLibrary.simpleMessage("Edit Profile"),
+    "email_must_not_empty": MessageLookupByLibrary.simpleMessage(
+      "Email must not empty",
+    ),
+    "enter_a_valid_email": MessageLookupByLibrary.simpleMessage(
+      "Enter a valid email",
+    ),
+    "enter_a_valid_password": MessageLookupByLibrary.simpleMessage(
+      "Oops! Your Password Is Not Correct",
+    ),
+    "enter_email": MessageLookupByLibrary.simpleMessage("Enter your Email"),
+    "enter_name": MessageLookupByLibrary.simpleMessage("Enter your Name"),
+    "enter_password": MessageLookupByLibrary.simpleMessage(
+      "Enter your Password",
+    ),
+    "enter_phone": MessageLookupByLibrary.simpleMessage("Enter your phone"),
+    "find_doctor": MessageLookupByLibrary.simpleMessage("Find\nDoctor"),
+    "find_doctors": MessageLookupByLibrary.simpleMessage("Find Doctors"),
+    "find_your_desire_health_solution": MessageLookupByLibrary.simpleMessage(
+      "Find your desire \nfor your health solution",
+    ),
+    "forgot_password": MessageLookupByLibrary.simpleMessage("Forgot password?"),
+    "go_to_home": MessageLookupByLibrary.simpleMessage("Go to home"),
+    "go_to_login_or_sign_up": MessageLookupByLibrary.simpleMessage(
+      "Go to Login or Signup",
+    ),
+    "height": MessageLookupByLibrary.simpleMessage("Height"),
+    "hi": MessageLookupByLibrary.simpleMessage("Hi"),
+    "language": MessageLookupByLibrary.simpleMessage("Language"),
+    "login": MessageLookupByLibrary.simpleMessage("Login"),
+    "logout": MessageLookupByLibrary.simpleMessage("Logout"),
+    "logoutGoogle": MessageLookupByLibrary.simpleMessage("Logout from google"),
+    "message": MessageLookupByLibrary.simpleMessage("Message"),
+    "messaging": MessageLookupByLibrary.simpleMessage("Messaging"),
+    "name_must_not_empty": MessageLookupByLibrary.simpleMessage(
+      "Name must not empty",
+    ),
+    "nurse": MessageLookupByLibrary.simpleMessage("Nurses"),
+    "once_again": MessageLookupByLibrary.simpleMessage(
+      "Once again you login successfully into medical app",
+    ),
+    "password_must_not_empty": MessageLookupByLibrary.simpleMessage(
+      "Password must not empty",
+    ),
+    "payment_detail": MessageLookupByLibrary.simpleMessage("Payment Detail"),
+    "payment_method": MessageLookupByLibrary.simpleMessage("Payment Method"),
+    "phone_must_not_empty": MessageLookupByLibrary.simpleMessage(
+      "phone must not empty",
+    ),
+    "please_check_your_Email": MessageLookupByLibrary.simpleMessage(
+      "Please Check your Email",
+    ),
+    "please_entre": MessageLookupByLibrary.simpleMessage(
+      "Please Entre E-mail\nit like gmail,yahoo,Hotmail,etc...",
+    ),
+    "producttype": MessageLookupByLibrary.simpleMessage("Product Type"),
+    "profile": MessageLookupByLibrary.simpleMessage("profile"),
+    "reason": MessageLookupByLibrary.simpleMessage("Reason"),
+    "recommendedtreatment": MessageLookupByLibrary.simpleMessage(
+      "Recommended Treatment",
+    ),
+    "save": MessageLookupByLibrary.simpleMessage("Save"),
+    "schedule": MessageLookupByLibrary.simpleMessage("Schedule"),
+    "search_your_doctor": MessageLookupByLibrary.simpleMessage(
+      "Search your doctor",
+    ),
+    "see_all": MessageLookupByLibrary.simpleMessage("Show all"),
+    "send_code_to_email": MessageLookupByLibrary.simpleMessage(
+      "Send code to Email",
+    ),
+    "sign_up": MessageLookupByLibrary.simpleMessage("Sign Up"),
+    "something_wrong_happen_please_Login_again":
+        MessageLookupByLibrary.simpleMessage(
+          "Something wrong happen Please Login again",
+        ),
+    "something_wrong_happen_please_sign_up_again":
+        MessageLookupByLibrary.simpleMessage(
+          "Something wrong happen Please Sign up again",
+        ),
+    "symptomsof": MessageLookupByLibrary.simpleMessage("symptoms"),
+    "top_doctor": MessageLookupByLibrary.simpleMessage("Top Doctor"),
+    "total": MessageLookupByLibrary.simpleMessage("Total"),
+    "upcoming": MessageLookupByLibrary.simpleMessage("Upcoming"),
+    "we_have_send_link_to_account_email": MessageLookupByLibrary.simpleMessage(
+      "we have send link to account Email",
+    ),
+    "weight": MessageLookupByLibrary.simpleMessage("Weight"),
+    "welcome_back": MessageLookupByLibrary.simpleMessage("Yeay! Welcome Back"),
+    "whenToConsultProfessional": MessageLookupByLibrary.simpleMessage(
+      "when To Consult Professional",
+    ),
+    "yes": MessageLookupByLibrary.simpleMessage("Yes"),
+    "you_have_account": MessageLookupByLibrary.simpleMessage(
+      "You have account? ",
+    ),
+    "you_have_already_signup_successfully_into_medidoc_app":
+        MessageLookupByLibrary.simpleMessage(
+          "You have already SignUp successfully into medidoc app",
+        ),
+  };
 }
