@@ -3,55 +3,56 @@ import 'package:flutter/material.dart';
 
 import '../../../app/style/custom_icon.dart';
 import '../../../const/colors.dart';
+import 'package:flutter/material.dart';
 
 class RowReusable extends StatelessWidget {
+  final VoidCallback? onTap;
+  final IconData leadingIcon;
+  final Color leadingIconColor;
+  final double leadingIconSize;
+  final Widget title;
+  final Widget? trailing;
+
   const RowReusable({
-    super.key,
+    Key? key,
     this.onTap,
     required this.leadingIcon,
-    this.leadingIconColor,
-    this.leadingIconSize,
-    this.title,
+    required this.leadingIconColor,
+    required this.leadingIconSize,
+    required this.title,
     this.trailing,
-  });
-  final void Function()? onTap;
-  final IconData leadingIcon;
-  final Color? leadingIconColor;
-  final double? leadingIconSize;
-  final Widget? title;
-  final Widget? trailing;
+  }) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return GestureDetector(
+    return InkWell(
       onTap: onTap,
-      child: Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.symmetric(
-              horizontal: 0,
-              vertical: 3,
-            ),
-            child: ListTile(
-              leading: CircleAvatar(
-                backgroundColor: AppColor.kGrey3,
-                child: CustomIcon(
-                  icon: leadingIcon,
-                  color: leadingIconColor,
-                  size: leadingIconSize,
-                ),
+      child: Container(
+        // ADD THIS MINIMUM HEIGHT/PADDING:
+        constraints: const BoxConstraints(minHeight: 56), // Material standard
+        padding: const EdgeInsets.symmetric(vertical: 14, horizontal: 0),
+        child: Row(
+          children: [
+            Container(
+              decoration: BoxDecoration(
+                color: leadingIconColor.withOpacity(0.12),
+                shape: BoxShape.circle,
               ),
-              trailing: trailing,
-              title: title,
+              padding: const EdgeInsets.all(8),
+              child: Icon(
+                leadingIcon,
+                color: leadingIconColor,
+                size: leadingIconSize,
+              ),
             ),
-          ),
-          Container(
-            //width: MediaQuery.of(context).size.width / 1.2,
-            height: 1.1,
-            decoration: const BoxDecoration(
-              color: AppColor.kGrey3,
-            ),
-          ),
-        ],
+            const SizedBox(width: 14),
+            Expanded(child: title),
+            if (trailing != null) ...[
+              const SizedBox(width: 8),
+              trailing!,
+            ],
+          ],
+        ),
       ),
     );
   }
